added eslint and refactored code
