+++ frontend/src/Components/Carousel/Carousel.css
@@ -1,0 +1,45 @@
+@import url('https://fonts.googleapis.com/css2?family=Bebas+Neue&family=IBM+Plex+Sans:wght@400&display=swap');
+
+.main-slider-container {
+  padding-left: 3rem;
+  padding-right: 3rem;
+  position: relative;
+  display: flex;
+  width: 100%;
+  margin-top: 1rem;
+  height: 12rem;
+  align-items: center;
+}
+
+#slider {
+  width: 100%;
+  height: 100%;
+  padding-top: 1rem;
+  white-space: nowrap;
+  overflow-x: auto;
+  scrollbar-width: none;
+  scroll-behavior: smooth;
+}
+
+#slider::-webkit-scrollbar {
+  display: none;
+}
+.slider-icon.left,
+.slider-icon.right {
+  background-color: white;
+  border-radius: 100%;
+  position: absolute;
+  opacity: 0.7;
+}
+.slider-icon.left {
+  left: 0;
+}
+.slider-icon.right {
+  right: 0;
+}
+.slider-icon.left:hover,
+.slider-icon.right:hover {
+  opacity: 1;
+  transform: scale(1.05);
+  transition: all ease-in-out 200ms;
+}
